+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - AzureMachinePool webhook handler that replaces mutators and validators.
 - AzureMachine webhook handler that replaces mutators and validators.
 - Spark webhook handler that replaces mutator.
+- AzureClusterConfig webhook handler that replaces validator.
 
 ### Changed